+++ src/window/listwindow.hpp
@@ -81,10 +81,10 @@ namespace Ui
       Main::WindowBuffer const & WindowBuffer() const { return lists_; }
 
    private:
-      void    SoftRedraw();
-      void    Clear();
-      size_t  BufferSize() const { return lists_.Size(); }
-      int32_t DetermineColour(uint32_t line) const;
+      void     SoftRedraw();
+      void     Clear();
+      uint32_t BufferSize() const { return lists_.Size(); }
+      int32_t  DetermineColour(uint32_t line) const;
 
    private:
       Main::Settings const & settings_;
